+++ com.unity.render-pipelines.universal/Runtime/UniversalRenderPipelineCore.cs
@@ -32,7 +32,6 @@ namespace UnityEngine.Rendering.Universal
         /// True if post-processing effect is enabled while rendering the camera stack.
         /// </summary>
         public bool postProcessingEnabled;
-        internal bool resolveFinalTarget;
     }
 
     [MovedFrom("UnityEngine.Rendering.LWRP")] public struct LightData
@@ -47,14 +46,51 @@ namespace UnityEngine.Rendering.Universal
 
     [MovedFrom("UnityEngine.Rendering.LWRP")] public struct CameraData
     {
+        // Internal camera data as we are not yet sure how to expose View in stereo context.
+        // We might change this API soon.
+        Matrix4x4 m_ViewMatrix;
+        Matrix4x4 m_ProjectionMatrix;
+
+        internal void SetViewAndProjectionMatrix(Matrix4x4 viewMatrix, Matrix4x4 projectionMatrix)
+        {
+            m_ViewMatrix = viewMatrix;
+            m_ProjectionMatrix = projectionMatrix;
+        }
+
+        /// <summary>
+        /// Returns the camera view matrix.
+        /// </summary>
+        /// <returns></returns>
+        public Matrix4x4 GetViewMatrix()
+        {
+            return m_ViewMatrix;
+        }
+
+        /// <summary>
+        /// Returns the camera projection matrix.
+        /// </summary>
+        /// <returns></returns>
+        public Matrix4x4 GetProjectionMatrix()
+        {
+            return m_ProjectionMatrix;
+        }
+
+        /// <summary>
+        /// Returns the camera GPU projection matrix. This contains platform specific changes to handle y-flip and reverse z.
+        /// Similar to <c>GL.GetGPUProjectionMatrix</c> but queries URP internal state to know if the pipeline is rendering to render texture. 
+        /// For more info on platform differences regarding camera projection check: https://docs.unity3d.com/Manual/SL-PlatformDifferences.html
+        /// </summary>
+        /// <seealso cref="GL.GetGPUProjectionMatrix(Matrix4x4, bool)"/>
+        /// <returns></returns>
+        public Matrix4x4 GetGPUProjectionMatrix()
+        {
+            return GL.GetGPUProjectionMatrix(m_ProjectionMatrix, IsCameraProjectionMatrixFlipped());
+        }
+
         public Camera camera;
         public CameraRenderType renderType;
         public RenderTexture targetTexture;
         public RenderTextureDescriptor cameraTargetDescriptor;
-        // Internal camera data as we are not yet sure how to expose View in stereo context.
-        // We might change this API soon.
-        internal Matrix4x4 viewMatrix;
-        internal Matrix4x4 projectionMatrix;
         internal Rect pixelRect;
         internal int pixelWidth;
         internal int pixelHeight;
@@ -66,6 +102,27 @@ namespace UnityEngine.Rendering.Universal
         public bool isHdrEnabled;
         public bool requiresDepthTexture;
         public bool requiresOpaqueTexture;
+
+        /// <summary>
+        /// True if the camera device projection matrix is flipped. This happens when the pipeline is rendering
+        /// to a render texture in non OpenGL platforms. If you are doing a custom Blit pass to copy camera textures
+        /// (_CameraColorTexture, _CameraDepthAttachment) you need to check this flag to know if you should flip the
+        /// matrix when rendering with for cmd.Draw* and reading from camera textures.
+        /// </summary>
+        public bool IsCameraProjectionMatrixFlipped()
+        {
+            // Users only have access to CameraData on URP rendering scope. The current renderer should never be null.
+            var renderer = ScriptableRenderer.current;
+            Debug.Assert(renderer != null, "IsCameraProjectionMatrixFlipped is being called outside camera rendering scope.");
+
+            if (renderer != null)
+            {
+                bool renderingToTexture = renderer.cameraColorTarget != BuiltinRenderTextureType.CameraTarget || targetTexture != null;
+                return SystemInfo.graphicsUVStartsAtTop && renderingToTexture;
+            }
+
+            return true;
+        }
 
         public SortingCriteria defaultOpaqueSortFlags;
 
@@ -91,6 +148,12 @@ namespace UnityEngine.Rendering.Universal
         public AntialiasingMode antialiasing;
         public AntialiasingQuality antialiasingQuality;
         internal ScriptableRenderer renderer;
+
+        /// <summary>
+        /// True if this camera is resolving rendering to the final camera render target.
+        /// When rendering a stack of cameras only the last camera in the stack will resolve to camera target.
+        /// </summary>
+        public bool resolveFinalTarget;
     }
 
     [MovedFrom("UnityEngine.Rendering.LWRP")] public struct ShadowData
@@ -160,6 +223,7 @@ namespace UnityEngine.Rendering.Universal
         public static readonly string DepthNoMsaa = "_DEPTH_NO_MSAA";
         public static readonly string DepthMsaa2 = "_DEPTH_MSAA_2";
         public static readonly string DepthMsaa4 = "_DEPTH_MSAA_4";
+        public static readonly string DepthMsaa8 = "_DEPTH_MSAA_8";
 
         public static readonly string LinearToSRGBConversion = "_LINEAR_TO_SRGB_CONVERSION";
         [Obsolete("The _KILL_ALPHA shader keyword is deprecated in the Universal Render Pipeline.")]